src/requirements.txt
PyArrow==11.0.0
pyspark==3.3.0
python_dateutil==2.8.2